+++ examples/example-protocol/src/main.rs
@@ -78,6 +78,7 @@ fp_import! {
     //
     // See `types/generics.rs` for more info.
     fn import_generics(arg: StructWithGenerics<u64>) -> StructWithGenerics<u64>;
+    fn import_explicit_bound_point(arg: ExplicitBoundPoint<u64>);
 
     // Custom type in a generic position.
     fn import_get_bytes() -> Result<ByteBuf, String>;